refactor : validate이름 변경 policy
